requirements.txt
asgiref==3.6.0
attrs==22.2.0
Django==3.2.18
django-cors-headers==3.14.0
django-extensions==3.2.1
django-filter==23.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-extra-fields==3.4.1
isort==5.11.5
Jinja2==3.1.2
MarkupSafe==2.1.2
oauthlib==3.2.2
packaging==23.0
Pillow==9.5.0
py==1.11.0
PyJWT==2.6.0
pytest==6.2.5
pytest-django==4.5.2
pytest-pythonpath==0.7.4
python3-openid==3.2.0
pytz==2023.3
reportlab==3.6.12
requests==2.28.2
requests-oauthlib==1.3.1
social-auth-app-django==4.0.0
social-auth-core==4.4.1
sqlparse==0.4.3
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15

